wds: allow profile reconfiguration
Extend uqmi so it can modify connection profiles stored on the modem.

This is required in case the operator disallows connections using the
dual-stacked PDP type. In case of either the home or visited network
disallowing dual-stacked connections, the PDP context establishment
fails with the default settings.

This can be worked around by configuring the APN and PDP type to profile
slot 1 and subsequently enabling the data connection.

Signed-off-by: David Bauer <mail@david-bauer.net>